reorganise files to folders
