+++ src/components/result/Result.js
