+++ src/components/search-data/SearchData.scss
