+++ src/components/search-input/SearchInput.scss
