+++ src/components/ui/spinner-overlay/Overlay.scss
